Kreirani i modifikovani kontroleri

+++ laravel-app/app/Http/Controllers/AuthorController.php
@@ -20,7 +20,18 @@ class AuthorController extends Controller
             'author'=>$author,
         ]);
     }
+    public function allauthors()
+    {
+        $author = Author::where('status','0')->get();
+        //return new AuthorCollection($authors);
 
+        return response()->json([
+            'status'=>200,
+            'author'=>$author,
+        ]);
+    }
+
+   
     public function show(Author $author)
     {
         return new AuthorResource($author);
@@ -36,9 +47,7 @@ class AuthorController extends Controller
         ]);
 
         if ($validator->fails()){ return response()->json(['status'=>400,'errors'=>$validator->messages()]);}
-        //$author = Author::create(['slug' => $request->slug, 'name' => $request->name, 'resting_place' => $request->resting_place]);
 
-        //return response()->json(['Author is created', new AuthorResource($author)]);
         else {
             $author = new Author;
             $author->slug=$request->input('slug');
@@ -51,7 +60,26 @@ class AuthorController extends Controller
         }
     }
 
-    public function update(Request $request, Author $author)
+    public function edit($id)
+{
+$author=Author::find($id);
+if($author)
+{
+return response()->json([
+'status'=>200,
+'author'=>$author
+]);
+}
+else
+{
+return response()->json([
+'status'=>404,
+'message'=>'No Author Id Found']);
+} }
+ 
+
+
+    public function update(Request $request, $id)
     {
         $validator = Validator::make($request->all(), [
             'slug' => 'required|string|max:100',
@@ -59,18 +87,60 @@ class AuthorController extends Controller
             'resting_place' => 'required|string'
         ]);
 
-        if ($validator->fails()) return response()->json($validator->errors());
-        $author->slug = $request->slug;
-        $author->name = $request->name;
-        $author->resting_place = $request->resting_place;
-        $author->save();
 
-        return response()->json(['Author is updated', new AuthorResource($author)]);
+        if($validator->fails())
+{
+return response()->json([
+'status'=>422,
+'errors'=>$validator->messages(),
+]); }
+else{
+
+    $author = Author::find($id);
+    
+    if($author) { 
+
+    $author->slug=$request->input('slug');
+    $author->name=$request->input('name');
+    $author->resting_place=$request->input('resting_place');
+    $author->status=$request->input('status')==true ? '1' : '0';
+    $author->save();
+    return response()->json([
+        'status'=>200,
+        'message'=>'Author updated successfully']);
+    }
+    else {
+        return response()->json([
+            'status'=>404,
+            'message'=>'No Author Id Found']);
+    
+
+
+
+
+    
+
+}
+}
     }
 
-    public function destroy(Author $author)
-    {
-        $author->delete();
-        return response()->json('Author is deleted');
-    }
+
+    public function destroy($id)
+{
+$author=Author::find($id);
+if($author)
+{
+$author->delete();
+return response()->json([
+'status'=>200,
+'message'=>'Author deleted successfully',
+]);
+}
+else{
+return response()->json([
+'status'=>404,
+'message'=>'No author ID found',
+]);
+}
+}
 }
